requirements.txt
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
GetOldTweets3==0.0.11
idna==2.10
lxml==4.6.3
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
Pillow==8.2.0
Pmw==2.0.1
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
tk==0.1.0
tkcalendar==1.6.1
tweepy==3.10.0
urllib3==1.26.5
